+++ x-pack/plugin/ml/src/test/java/org/elasticsearch/xpack/ml/job/results/ReservedFieldNamesTests.java
@@ -11,11 +11,10 @@ import org.elasticsearch.xpack.core.ml.job.results.ReservedFieldNames;
 
 public class ReservedFieldNamesTests extends ESTestCase {
 
-    public void testIsValidFieldName() throws Exception {
+    public void testIsValidFieldName() {
         assertTrue(ReservedFieldNames.isValidFieldName("host"));
         assertTrue(ReservedFieldNames.isValidFieldName("host.actual"));
         assertFalse(ReservedFieldNames.isValidFieldName("actual.host"));
         assertFalse(ReservedFieldNames.isValidFieldName(AnomalyRecord.BUCKET_SPAN.getPreferredName()));
     }
-
 }
